Implement supersampling for 3D previews
This greatly improves the appearance of 3D previews by reducing
specular and parallax aliasing.

It has a significant performance cost, so it can be disabled in
the preferences by setting the resolution scale to 1.0.

The default setting is 2.0 (2× supersampling), but it can be increased
up to 2.5 in the preferences for even better results.

The 3D viewport's MSAA has been decreased to 2× to account for this
change.