Decrease the maximum allowed value for 3D preview resolution to 2.0
Values above 2.0 actually *decrease* rendering quality due to
the renderer not having any mipmapped images of the viewport
to work with.

It might be possible to use some kind of real-time mipmap generation,
but it would be very slow and would not benefit the output quality much
(4× SSAA is pretty good already).