New Theme "Modern"
The theme resource was done from scratch to avoid the use of unnecessary/outdated theme properties (a clean theme resource should make it easier to adjust and vary in the future).

Things don't really look perfectly like I envision them, because there are some panels and other customization options missing. But I think this is a good start for a new theme. I've not made it the new default, just added it to the theme list.

The theme uses the Rubik font, as suggested by Tarox.