Implement SVG Icons (to remove icon folders)
Adds in an SVG icon sheet. This allows having color-changing icons and possibly having scalable icons in the future as well.

This required changing the icons in some places. We also don't want to manually reference the icon-sheet in scenes, but instead always use the theme (add the icon to the theme, e.g. the MM_Icons type) then assign them in code with `get_theme_icon("settings", "MM_Icons")`. In some cases we might have to update this on theme changes listening to NOTIFICATION_THEME_CHANGED.