Split Default theme into base and variations
We now have a default.tres theme and both default dark and default light are using the enhanced theme system. This makes them behave more similar. The default is still using the same base colors as dark, this could be changed in the future, probably when we want to expose the colors to the users.